pixela-java-client-75 separate Klass/Choice/Choices/Range into independent classes.
