+++ api/src/test/java/pixela/client/Range.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2019 Shinya Mochida
+ *
+ * Licensed under the Apache License,Version2.0(the"License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,software
+ * Distributed under the License is distributed on an"AS IS"BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package pixela.client;
+
+class Range {
+  final int min;
+  final int size;
+
+  Range(final int min, final int size) {
+    this.min = min;
+    this.size = size;
+  }
+
+  @Override
+  public String toString() {
+    return "range[min:" + min + ",size:" + size + "]";
+  }
+}
